refactored : host_interface
