+++ libp2p/host/ping.py
@@ -4,11 +4,11 @@ import time
 
 import trio
 
+from libp2p.abc import (
+    IHost,
+)
 from libp2p.custom_types import (
     TProtocol,
-)
-from libp2p.host.host_interface import (
-    IHost,
 )
 from libp2p.network.stream.exceptions import (
     StreamClosed,
