+++ libp2p/identity/identify/protocol.py
@@ -4,12 +4,12 @@ from multiaddr import (
     Multiaddr,
 )
 
+from libp2p.abc import (
+    IHost,
+)
 from libp2p.custom_types import (
     StreamHandlerFn,
     TProtocol,
-)
-from libp2p.host.host_interface import (
-    IHost,
 )
 from libp2p.network.stream.exceptions import (
     StreamClosed,
