+++ libp2p/tools/pubsub/dummy_account_node.py
@@ -6,7 +6,7 @@ from contextlib import (
     asynccontextmanager,
 )
 
-from libp2p.host.host_interface import (
+from libp2p.abc import (
     IHost,
 )
 from libp2p.pubsub.pubsub import (
